fix: 修复了解析json数据类型错误的bug

+++ src/main/kotlin/cn/rtast/rmusic/models/netease/detail/M.kt
@@ -7,9 +7,9 @@
 package cn.rtast.rmusic.models.netease.detail
 
 data class M(
-    val br: Int,
-    val fid: Int,
-    val size: Int,
-    val sr: Int,
-    val vd: Int
+    val br: Long,
+    val fid: Long,
+    val size: Long,
+    val sr: Long,
+    val vd: Long
 )
