+++ src/main/kotlin/cn/rtast/rmusic/models/netease/login/Account.kt
@@ -8,18 +8,18 @@ package cn.rtast.rmusic.models.netease.login
 
 data class Account(
     val anonimousUser: Boolean,
-    val ban: Int,
-    val baoyueVersion: Int,
+    val ban: Long,
+    val baoyueVersion: Long,
     val createTime: Long,
-    val donateVersion: Int,
+    val donateVersion: Long,
     val id: Long,
     val salt: String,
-    val status: Int,
-    val tokenVersion: Int,
-    val type: Int,
+    val status: Long,
+    val tokenVersion: Long,
+    val type: Long,
     val uninitialized: Boolean,
     val userName: String,
-    val vipType: Int,
+    val vipType: Long,
     val viptypeVersion: Long,
-    val whitelistAuthority: Int
+    val whitelistAuthority: Long
 )
