+++ src/main/kotlin/cn/rtast/rmusic/models/netease/login/AvatarDetail.kt
@@ -8,6 +8,6 @@ package cn.rtast.rmusic.models.netease.login
 
 data class AvatarDetail(
     val identityIconUrl: String,
-    val identityLevel: Int,
-    val userType: Int
+    val identityLevel: Long,
+    val userType: Long
 )
