+++ src/main/kotlin/cn/rtast/rmusic/models/netease/login/Binding.kt
@@ -9,11 +9,11 @@ package cn.rtast.rmusic.models.netease.login
 data class Binding(
     val bindingTime: Long,
     val expired: Boolean,
-    val expiresIn: Int,
+    val expiresIn: Long,
     val id: Long,
-    val refreshTime: Int,
+    val refreshTime: Long,
     val tokenJsonStr: String,
-    val type: Int,
+    val type: Long,
     val url: String,
     val userId: Long
 )
